requirements.txt
Flask
gunicorn
python-dotenv
matplotlib
numpy
pandas
Pillow
psycopg[binary]
reportlab
requests
